requirements.txt
asgiref==3.5.2
automium==0.2.6
automium-web==0.1.1
certifi==2022.9.24
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
Django==4.1.3
djangorestframework==3.14.0
drf-yasg==1.21.4
Flask==2.2.2
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
legacy==0.1.6
MarkupSafe==2.1.1
packaging==21.3
pyparsing==3.0.9
pytz==2022.6
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.13
Werkzeug==2.2.2
pip==22.3.1